requirements.txt
pynput
pytesseract
google-generativeai
python-dotenv
Pillow
plyer
mss
pywin32
pyautogui
opencv-python
